requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
zipp==3.23.0
